+++ src/ffpopt/geom/GeomOptAse.py
@@ -108,8 +108,9 @@ def GeomOpt_ASE(los,struct,constraints=None,restraints=None):
     
     """
     import os
-    if "OMP_NUM_THREADS" not in os.environ:
-        os.environ["OMP_NUM_THREADS"] = "1"
+    from ffpopt.runtime.CpuThreads import pin_math_threads
+
+    pin_math_threads(1)
  
     import copy
     import sys